created db package and class, deleted unnecessary rows

+++ src/university/research/Researcher.java
@@ -4,9 +4,7 @@ package university.research;
 public  interface Researcher 
 {
 
-	
-	 int calculateHIndex() ;
-	
+	 double calculateHIndex() ;
 
 	 ResearchPaper printPapers(ResearchPaper parameter) ;
